fastbootd: Allow using kernel commits to pass SafetyNet
When using kernel commits to pass SafetyNet and CTS like:

  proc: Remove verifiedbootstate flag from /proc/cmdline
  proc: Set androidboot.verifiedbootstate=green

fastbootd returns the device as locked when flashing images
or resizing partitions when flashing because it can't find
androidboot.verifiedbootstate=orange in the kernel command
line letting it know it's safe to flash.

So if androidboot.verifiedbootstate is missing from the kernel
command line or if is returning verifiedbootstate=green report
the device as unlocked.

PS2: Add support for androidboot.verifiedbootstate=green

Change-Id: I56712f374ea26397ea5a183cfe100cf92e21d531
Signed-off-by: Simão Gomes Viana <devel@superboring.dev>